CST-FROM-EXPRESSION does not create table entries for value-like atoms
Before and after this change, if the EXPRESSION->CST table says that a
particular CST should be used for a given atom, then that CST is
used.

However, for a "value-like" atom (number, character, symbol) that has
no matching entry in the EXPRESSION->CST table, the new and old
behavior differs: Before this change, a CST was created and an entry
for the atom and the CST was added to the table. The input expression
can contain "other" atoms that are EQL to the atom in question and
since EXPRESSION->CST usually uses EQ or EQL as its test, those
instances would be represented by the same CST in the overall result
CST. After this change, each of the "other" atoms is represented by a
fresh CST.

Example:

Input
  (1 1 :foo :foo #1=#P"foo" #1#)

Before this change
  (eq (cst:first result) (cst:second result))
  (eq (cst:third result) (cst:fourth result))
  (eq (cst:fifth result) (cst:sixth result))

After this change
  (not (eq (cst:first result) (cst:second result)))
  (not (eq (cst:third result) (cst:fourth result)))
  (eq (cst:fourth result) (cst:sixth result))

The old behavior was roughly equivalent to treating the input
expression as

  (#1=1 #1# #2=:foo #2# #3=#P"foo" #3#)

which does not make sense since the question of whether occurrences of
atom are the "same" instance of that atom does not make sense for
numbers, characters and symbols.